restructures
